backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow>=10.0.0
numpy>=1.24.0
timm>=0.9.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
